requirements.txt
Django==3.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
dj-rest-auth==1.1.1
django-allauth==0.42.0
djangorestframework==3.11.1

